+++ fileserver/src/main/java/org/jugs/webdav/fileserver/tools/PropStatBuilderExt.java
@@ -27,14 +27,14 @@ import java.util.Set;
 
 import javax.ws.rs.core.Response.Status;
 
-import net.java.dev.webdav.jaxrs.xml.elements.Prop;
-import net.java.dev.webdav.jaxrs.xml.elements.PropStat;
-import net.java.dev.webdav.jaxrs.xml.properties.CreationDate;
-import net.java.dev.webdav.jaxrs.xml.properties.DisplayName;
-import net.java.dev.webdav.jaxrs.xml.properties.GetContentLength;
-import net.java.dev.webdav.jaxrs.xml.properties.GetContentType;
-import net.java.dev.webdav.jaxrs.xml.properties.GetLastModified;
-import net.java.dev.webdav.jaxrs.xml.properties.ResourceType;
+import org.jugs.webdav.jaxrs.xml.elements.Prop;
+import org.jugs.webdav.jaxrs.xml.elements.PropStat;
+import org.jugs.webdav.jaxrs.xml.properties.CreationDate;
+import org.jugs.webdav.jaxrs.xml.properties.DisplayName;
+import org.jugs.webdav.jaxrs.xml.properties.GetContentLength;
+import org.jugs.webdav.jaxrs.xml.properties.GetContentType;
+import org.jugs.webdav.jaxrs.xml.properties.GetLastModified;
+import org.jugs.webdav.jaxrs.xml.properties.ResourceType;
 
 import org.w3c.dom.Element;
 
@@ -152,7 +152,7 @@ public class PropStatBuilderExt {
 		if(!empty){
 			Object[] objects = notFound.toArray(new Object[properties.size()]);
 			Prop prop = new Prop(objects);
-			stat = new PropStat(prop, new net.java.dev.webdav.jaxrs.xml.elements.Status(Status.NOT_FOUND));
+			stat = new PropStat(prop, new org.jugs.webdav.jaxrs.xml.elements.Status(Status.NOT_FOUND));
 		}		
 		
 		return stat;
@@ -167,7 +167,7 @@ public class PropStatBuilderExt {
 	public PropStat build(){
 		Object[] objects = properties.toArray(new Object[properties.size()]);
 		Prop prop = new Prop(objects);
-		PropStat stat = new PropStat(prop, new net.java.dev.webdav.jaxrs.xml.elements.Status(status));
+		PropStat stat = new PropStat(prop, new org.jugs.webdav.jaxrs.xml.elements.Status(status));
 		
 		return stat;
 	}